added cyclicbuffer.h header file
This file is the required header file for the Arduino program to collect data. This is used to record sensor data after the sensor input is detected. CyclicBuffer constantly records the sensor data in an array that can record up to a defined amount of values. When an event occurs it will get previous data, then the rest of the program will combine the real-time data and cyclicbuffer data into a single array.