added the Arduino file
This program is to record the data from the sensors. It will use a CyclicBuffer to records previous data and when an event occurs it will record real-time data and after the recording it will combine both buffer data and real-time data in to a single array and pass it using serial communication to the computer.